Add project directory structure and preliminary server code
This commit lays out the directory structure for the project,
as well as adding several methods and routes for the server.
Included is a listener to receive HTTP post requests which
will be used to receive and store data.

Much of this server code was 'borrowed' from haus at
https://github.com/haus/metrics-server (a fork may have been
more sane).